% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{add_definite_article_de}
\alias{add_definite_article_de}
\title{Add the German definite article in dative case to a preposition}
\usage{
add_definite_article_de(
  preposition,
  gender = c("feminine", "masculine", "neuter")
)
}
\arguments{
\item{preposition}{The preposition to add the German definite article in dative case to. A character scalar.}

\item{gender}{The grammatical gender of the \code{preposition}.}
}
\value{
A character scalar.
}
\description{
Add the German definite article in dative case to a preposition
}
\seealso{
Other spoken language functions:
\code{\link{decline_noun_de}()},
\code{\link{definite_article_de}()},
\code{\link{definite_article_de_declined}()},
\code{\link{n_ify}()},
\code{\link{phrase_datetime}()},
\code{\link{prettify_date}()},
\code{\link{prettify_datetime}()},
\code{\link{rank_nr}()},
\code{\link{write_out_n}()}
}
\concept{spoken}
