% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim-GEN.R
\name{definite_article_de}
\alias{definite_article_de}
\title{Get the German definite article}
\usage{
definite_article_de(
  gender = c("feminine", "masculine", "neuter"),
  case = c("nominative", "genitive", "dative", "accusative"),
  pluralize = FALSE
)
}
\arguments{
\item{gender}{The grammatical gender of the subject the definite article refers to.}

\item{case}{The grammatical case of the subject the definite article refers to.}

\item{pluralize}{Whether to return the plural form of the definite article. A logical scalar.}
}
\value{
A character scalar.
}
\description{
Get the German definite article
}
\seealso{
Other spoken language functions:
\code{\link{add_definite_article_de}()},
\code{\link{decline_noun_de}()},
\code{\link{definite_article_de_declined}()},
\code{\link{n_ify}()},
\code{\link{rank_nr}()},
\code{\link{write_out_n}()}
}
\concept{spoken}
