% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim-GEN.R
\name{definite_article_de_declined}
\alias{definite_article_de_declined}
\title{Decline the German definite article}
\usage{
definite_article_de_declined(
  gender = c("feminine", "masculine", "neuter"),
  case = c("nominative", "genitive", "dative", "accusative"),
  pluralize = FALSE
)
}
\arguments{
\item{gender}{The grammatical gender of the subject the definite article refers to.}

\item{case}{The grammatical case of the subject the definite article refers to.}

\item{pluralize}{Whether to return the plural form of the definite article. A logical scalar.}
}
\value{
A character scalar.
}
\description{
The declination of the German definite article \href{https://de.wikipedia.org/wiki/Relativpronomen#Der,_die,_das}{is a relative pronoun}.
}
\details{
Note that the second singular feminine and plural genitive form \code{"derer"} is ignored by this function.
}
\seealso{
Other spoken language functions:
\code{\link{add_definite_article_de}()},
\code{\link{decline_noun_de}()},
\code{\link{definite_article_de}()},
\code{\link{n_ify}()},
\code{\link{rank_nr}()},
\code{\link{write_out_n}()}
}
\concept{spoken}
