% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{download_pandoc_binaries}
\alias{download_pandoc_binaries}
\title{Download Pandoc release}
\usage{
download_pandoc_binaries(
  release_id = yay::gh_release_latest(owner = "jgm", name = "pandoc")$id,
  os = c("linux", "macos", "windows"),
  path = "bin/",
  overwrite = TRUE
)
}
\arguments{
\item{release_id}{The GitHub release identifier of the desired Pandoc release. Use
\code{\link[yay:gh_releases]{yay::gh_releases(owner = "jgm", name = "pandoc")}} to determine the release identifier of a specific Pandoc version number. An
integer scalar.}

\item{os}{The operating system(s) for which Pandoc binaries should be downloaded. Any combination of
\code{"linux"}, \code{"macos"} and \code{"windows"}.}

\item{path}{The filesystem path to which the binaries are saved to. A \link[fs:fs_path]{path} or something coercible to.}

\item{overwrite}{Whether to overwrite existing binaries under \code{path}.}
}
\value{
\code{path} invisibly.
}
\description{
Downloads the \link[=pandoc_release_assets]{assets} and extracts the executable program binaries for Linux, macOS and Windows of the specified Pandoc release.
}
\seealso{
GitHub release functions: \code{\link[yay:gh_releases]{yay::gh_releases()}}, \code{\link[yay:gh_release_latest]{yay::gh_release_latest()}}

Other Pandoc functions:
\code{\link{pandoc_release_assets}()},
\code{\link{pandoc_tpl}()}
}
\concept{pandoc}
