% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{gen_quarto_snips}
\alias{gen_quarto_snips}
\title{Generate Quarto knitr code chunks from structured data}
\usage{
gen_quarto_snips(data, path, type = c("fig", "tbl"))
}
\arguments{
\item{data}{Input data as returned by \code{\link[=read_quarto_chunk_data]{read_quarto_chunk_data()}}. A dataframe where column names correspond to chunk options supported by
\code{\link[=quarto_fig_chunk]{quarto_fig_chunk()}} or \code{\link[=quarto_tbl_chunk]{quarto_tbl_chunk()}}, depending on \code{type}. Columns \code{body}, \code{label} and \code{fig_cap}/\code{tbl_cap} are mandatory.}

\item{path}{Output directory where the generated figure chunk \code{.qmd} files are to be saved.}

\item{type}{Quarto code chunk type. One of \code{"fig"} or \code{"tbl"}.}
}
\value{
\code{data}, invisibly.
}
\description{
Generates Quarto knitr \link[=quarto_fig_chunk]{figure} or \link[=quarto_tbl_chunk]{table} code chunks from structured \code{data}.
}
\details{
The column names of \code{data} can either correspond to \code{\link[=quarto_fig_chunk]{quarto_fig_chunk()}}/\code{\link[=quarto_tbl_chunk]{quarto_tbl_chunk()}}'s argument names (in \href{https://en.wikipedia.org/wiki/Snake_case}{snake case}), to \href{https://yihui.org/knitr/options/#plots}{knitr's figure chunk option names} (period-separared) or
directly to Quarto's
\href{https://quarto.org/docs/reference/cells/cells-knitr.html#figures}{figure}/\href{https://quarto.org/docs/reference/cells/cells-knitr.html#tables}{table} code
chunk option names (in \href{https://en.wikipedia.org/wiki/Letter_case#Kebab_case}{kebab case}), they will be converted as needed. Additional column names are
silently ignored.

The generated \code{.qmd} files are intended to be included in other Quarto documents via the built-in
\href{https://quarto.org/docs/authoring/includes.html}{\code{include}} shortcode.
}
\examples{
tmp_path <- fs::path_temp("example-gen_quarto_snips") |> fs::dir_create()

tibble::tibble(label = "fig-cars",
               fig_cap = "Today's ({lubridate::today()}) top figure",
               fig_column = "page",
               fig_height = 8L,
               body = "plot(cars)") |>
  tibble::add_row(label = "fig-mtcars",
                  fig_cap = "More cars",
                  fig_column = "screen",
                  fig_height = 8L,
                  body = "plot(mtcars)") |>
  salim::gen_quarto_snips(type = "fig",
                          path = tmp_path)

fs::dir_delete(tmp_path)
}
\seealso{
Other Quarto functions:
\code{\link{quarto_fig_chunk}()},
\code{\link{quarto_tbl_chunk}()},
\code{\link{read_quarto_chunk_data}()}
}
\concept{quarto}
