% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{ggplot2_geom_defaults}
\alias{ggplot2_geom_defaults}
\title{Set common ggplot2 geom defaults}
\usage{
ggplot2_geom_defaults(family)
}
\arguments{
\item{family}{Font family to use in text geoms.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Sets common ggplot2 geom default settings (currently only font \code{family}). Convenience wrapper around \code{\link[ggplot2:update_defaults]{ggplot2::update_geom_defaults()}}.
}
\details{
Note that \code{ggplot2_geom_defaults()} must be invoked \emph{before} plots are created to have any effect.
}
\seealso{
Other plot theming functions:
\code{\link{ggplot2_theme}()},
\code{\link{ggplot2_theme_html}()},
\code{\link{plotly_layout}()}
}
\concept{plot_theme}
