% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{ggplot2_theme}
\alias{ggplot2_theme}
\title{Apply common theme to ggplot2 chart}
\usage{
ggplot2_theme(
  base_size = 11L,
  base_family = pal::pkg_config_val(key = "font_family_body", pkg = this_pkg),
  base_line_size = base_size/22L,
  base_rect_size = base_size/22L,
  axis.title.x = ggplot2::element_text(margin = ggplot2::margin(t = 20L, r = 0L, b = 0L,
    l = 0L), inherit.blank = TRUE),
  axis.title.y = ggplot2::element_text(margin = ggplot2::margin(t = 0L, r = 20L, b = 0L,
    l = 0L), inherit.blank = TRUE),
  legend.box.margin = ggplot2::margin(),
  legend.box.spacing = ggplot2::element_blank(),
  legend.margin = ggplot2::margin(),
  legend.position = "bottom",
  legend.spacing = grid::unit(x = 0, units = "npc"),
  plot.margin = ggplot2::margin(),
  ...
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{axis.title.x, axis.title.y}{Labels of axes (\code{\link[ggplot2:element]{ggplot2::element_text()}}). Specify all axes' labels (\code{axis.title}), labels by plane (using \code{axis.title.x}
or \code{axis.title.y}), or individually for each axis (using \code{axis.title.x.bottom}, \code{axis.title.x.top}, \code{axis.title.y.left}, \code{axis.title.y.right}).
\verb{axis.title.*.*} inherits from \verb{axis.title.*} which inherits from \code{axis.title}, which in turn inherits from \code{text}.}

\item{legend.box.margin}{margins around the full legend area, as specified
using \code{\link[ggplot2:margin]{margin()}}}

\item{legend.box.spacing}{The spacing between the plotting area and the
legend box (\code{unit})}

\item{legend.margin}{the margin around each legend (\code{\link[ggplot2:margin]{margin()}})}

\item{legend.position}{the default position of legends ("none", "left",
"right", "bottom", "top", "inside")}

\item{legend.spacing}{Spacing between legends (\code{unit}). \code{legend.spacing.x} & \code{legend.spacing.y} inherit from \code{legend.spacing} or can be specified separately.}

\item{plot.margin}{margin around entire plot (\code{unit} with the sizes of
the top, right, bottom, and left margins)}

\item{...}{Further arguments passed on to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
An object of class \code{\link[ggplot2:theme]{theme}}.
}
\description{
Applies common \link[ggplot2:theme]{theming} to a \link[ggplot2:ggplot2-package]{ggplot2} chart. It's based on \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}.
}
\examples{
sysfonts::font_add_google(name = "Alegreya Sans")
showtext::showtext_auto()

ggplot2::ggplot(data = mtcars,
                mapping = ggplot2::aes(x = mpg,
                                       y = cyl)) +
  ggplot2::geom_point() +
  salim::ggplot2_theme(base_size = 12)
}
\seealso{
Other plot theming functions:
\code{\link{ggplot2_geom_defaults}()},
\code{\link{ggplot2_theme_html}()},
\code{\link{plotly_layout}()}
}
\concept{plot_theme}
