% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{lvl_up_quarto}
\alias{lvl_up_quarto}
\title{Level up Quarto}
\usage{
lvl_up_quarto(path_min_vrsn, update_min_vrsn = FALSE)
}
\arguments{
\item{path_min_vrsn}{Path to the cached Quarto version string.}

\item{update_min_vrsn}{Whether or not to overwrite \code{path_min_vrsn} with the currently installed Quarto version string \emph{iff} the latter is higher than the
former.}
}
\value{
Currently installed Quarto version as a \link[=numeric_version]{numeric version}, invisibly.
}
\description{
Checks whether the installed \href{https://quarto.org/}{Quarto} version is >= a cached reference version. Intended to level up the Quarto version in use between
multiple users of the same code (e.g. contributors to a specific Quarto project).
}
\seealso{
Other development environment currentness functions:
\code{\link{lvl_up_r}()},
\code{\link{lvl_up_rstudio}()},
\code{\link{update_rpkgs}()},
\code{\link{update_salims_pkgs}()},
\code{\link{update_zdaarau_pkgs}()}
}
\concept{dev_env}
