% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{lvl_up_r}
\alias{lvl_up_r}
\title{Level up R}
\usage{
lvl_up_r(path_min_vrsn, update_min_vrsn = FALSE)
}
\arguments{
\item{path_min_vrsn}{Path to the cached R version string.}

\item{update_min_vrsn}{Whether or not to overwrite \code{path_min_vrsn} with the currently installed R version string \emph{iff} the latter is higher than the former.}
}
\value{
Currently installed R version as a \link[=as.package_version]{numeric version}, invisibly.
}
\description{
Checks whether the installed R version is >= a cached reference version. Intended to level up the R version in use between multiple users of the same code
(e.g. contributors to a specific R project).
}
\details{
If the file \code{path_min_vrsn} exists, it is checked whether the installed R version is greater than or equal to the version number stored in that file, and if
not, an alert is displayed.

If \code{update_min_vrsn} is set to \code{TRUE} and the file \code{path_min_vrsn} doesn't exist or contains an R version string that's lower than the currently installed
version of R, \code{path_min_vrsn} is overwritten with the currently installed R version string.
}
\seealso{
Other development environment currentness functions:
\code{\link{lvl_up_quarto}()},
\code{\link{lvl_up_rstudio}()},
\code{\link{update_salims_pkgs}()}
}
\concept{dev_env}
