% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{n_ify}
\alias{n_ify}
\title{Append \code{"n"} to indefinite article where indicated}
\usage{
n_ify(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
Appends an \code{"n"} to the indefinite article "a" if required by English language rules.
}
\details{
Note that this function only covers part of all cases where "an" instead of "a" is indicated. Since the choice of \emph{a} or \emph{an} \href{http://www.butte.edu/departments/cas/tipsheets/grammar/articles.html}{depends on sound, not spelling}, it's not trivial to implement a universal solution.

For example, this function doesn't properly append an \code{"n"} to the indefinite article in the following situations:
\itemize{
\item A word starting with the letter "u" where it is pronounced as a vowel, e.g. \emph{\strong{unhappy}}. Note that most English words starting with letter "u" are
pronounced differently like \emph{universum} and thus require the indefinite article \emph{a}, not \emph{an}.
\item A word starting with a consonant that is pronounced as a vowel, e.g. \emph{\strong{hour}}.
}
}
\examples{
n_ify("This is a absolute disaster, such a error.")
}
\seealso{
Other spoken language functions:
\code{\link{add_definite_article_de}()},
\code{\link{decline_noun_de}()},
\code{\link{definite_article_de_declined}()},
\code{\link{definite_article_de}()},
\code{\link{rank_nr}()},
\code{\link{write_out_n}()}
}
\concept{spoken}
