% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{pandoc_release_assets}
\alias{pandoc_release_assets}
\title{List Pandoc release assets}
\usage{
pandoc_release_assets(
  release_id = yay::gh_release_latest(owner = "jgm", name = "pandoc")$id
)
}
\arguments{
\item{release_id}{The GitHub release identifier of the desired Pandoc release. Use
\code{\link[yay:gh_releases]{yay::gh_releases(owner = "jgm", name = "pandoc")}} to determine the release identifier of a specific Pandoc version number. An
integer scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Uses \code{\link[gh:gh]{gh::gh()}} to fetch filenames, corresponding operating systems and download URLs of a specific \href{https://docs.github.com/repositories/releasing-projects-on-github}{GitHub release} of \href{https://pandoc.org/}{Pandoc} via \href{https://docs.github.com/en/rest/reference/repos#list-release-assets}{GitHub's REST API} and returns them as a \link[tibble:tbl_df-class]{tibble}.
}
\seealso{
GitHub release functions: \code{\link[yay:gh_releases]{yay::gh_releases()}}, \code{\link[yay:gh_release_latest]{yay::gh_release_latest()}}

Other Pandoc functions:
\code{\link{download_pandoc_binaries}()},
\code{\link{pandoc_tpl}()}
}
\concept{pandoc}
