% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{pandoc_release_id_latest}
\alias{pandoc_release_id_latest}
\title{Get latest Pandoc release ID}
\usage{
pandoc_release_id_latest()
}
\value{
An integer scalar.
}
\description{
Uses \code{\link[gh:gh]{gh::gh()}} to fetch the latest \href{https://pandoc.org/}{Pandoc} release ID via \href{https://docs.github.com/en/rest/reference/repos#get-the-latest-release}{GitHub's REST API}.
}
