% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{pandoc_tpl}
\alias{pandoc_tpl}
\title{Download Pandoc template}
\usage{
pandoc_tpl(tpl = "quarto_mod.latex")
}
\arguments{
\item{tpl}{Template to download. One of \code{"quarto_mod.latex"}.}
}
\value{
The template as a character scalar.
}
\description{
Downloads one of our \href{https://gitlab.com/salim_b/pandoc/templates}{custom Pandoc templates} and returns it as a character scalar. The template is cached
using \code{\link[pal:req_cached]{pal::req_cached()}}, thus only re-downloaded if the local copy is outdated.
}
\examples{
\dontrun{
salim::pandoc_tpl(tpl = "quarto_mod.latex") |>
  brio::write_file(path = "quarto_mod.latex")}
}
\seealso{
Other Pandoc functions:
\code{\link{download_pandoc_binaries}()},
\code{\link{pandoc_release_assets}()},
\code{\link{pandoc_release_id_latest}()},
\code{\link{pandoc_releases}()},
\code{\link{pandoc_version_latest}()}
}
\concept{pandoc}
