% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{pandoc_version_latest}
\alias{pandoc_version_latest}
\title{Get latest Pandoc release version number}
\usage{
pandoc_version_latest()
}
\value{
A \link[=numeric_version]{numeric version}.
}
\description{
Uses \code{\link[gh:gh]{gh::gh()}} to fetch \href{https://pandoc.org/}{Pandoc}'s latest \href{https://docs.github.com/repositories/releasing-projects-on-github}{GitHub release} version
number via \href{https://docs.github.com/en/rest/reference/repos#get-the-latest-release}{GitHub's REST API} and returns it as a \link[=numeric_version]{numeric version}.
}
\seealso{
Other Pandoc functions:
\code{\link{download_pandoc_binaries}()},
\code{\link{pandoc_release_assets}()},
\code{\link{pandoc_release_id_latest}()},
\code{\link{pandoc_releases}()},
\code{\link{pandoc_tpl}()}
}
\concept{pandoc}
