% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{plotly_layout}
\alias{plotly_layout}
\title{Apply common layout to Plotly chart}
\usage{
plotly_layout(
  p,
  font = list(color = pal::pkg_config_val("plot_color_body"), family =
    pal::pkg_config_val("font_family_body")),
  paper_bgcolor = pal::pkg_config_val("plot_color_bg"),
  plot_bgcolor = paper_bgcolor,
  xaxis = list(gridcolor = pal::pkg_config_val("plot_color_grid")),
  yaxis = list(gridcolor = pal::pkg_config_val("plot_color_grid")),
  ...
)
}
\arguments{
\item{p}{\link[plotly:plot_ly]{Plotly object} to modify.}

\item{font}{\href{https://plotly.com/r/reference/layout/#layout-font}{Global font}. Fonts used in traces and other layout components inherit from the global font.}

\item{paper_bgcolor}{\href{https://plotly.com/r/reference/layout/#layout-paper_bgcolor}{Background color of the paper} where the graph is drawn.}

\item{plot_bgcolor}{\href{https://plotly.com/r/reference/layout/#layout-plot_bgcolor}{Background color of the plotting area} in-between x and y axes.}

\item{xaxis}{\href{https://plotly.com/r/reference/layout/#layout-xaxis}{X-axis configuration}.}

\item{yaxis}{\href{https://plotly.com/r/reference/layout/#layout-yaxis}{Y-axis configuration}.}

\item{...}{Further \href{https://plotly.com/r/reference/layout/}{layout configuration} arguments.}
}
\value{
A \link[plotly:plot_ly]{plotly object}.
}
\description{
Applies common \link[plotly:layout]{layout} configuration to a \link[plotly:plot_ly]{Plotly} chart.
}
\examples{
plotly::plot_ly(data = mtcars,
                type = "bar",
                x = ~mpg) |>
  salim::plotly_layout(font = list(color = pal::pkg_config_val(key = "plot_color_body",
                                                               pkg = "salim")))
}
\seealso{
Other plot theming functions:
\code{\link{ggplot2_geom_defaults}()},
\code{\link{ggplot2_theme}()},
\code{\link{ggplot2_theme_html}()}
}
\concept{plot_theme}
