% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{quarto_tbl_chunk}
\alias{quarto_tbl_chunk}
\title{Assemble Quarto knitr table chunk}
\usage{
quarto_tbl_chunk(
  body,
  label,
  tbl_cap,
  tbl_subcap = NULL,
  layout_class = "body"
)
}
\arguments{
\item{body}{\R code to insert into the code chunk's body. A character scalar.}

\item{label}{Unique code chunk label. Set as Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#tables}{\code{label}} code chunk option. A
character scalar that starts with \code{"tbl-"}.}

\item{tbl_cap}{Table caption. Set as Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#tables}{\code{tbl-cap}} code chunk option. A character
scalar.}

\item{tbl_subcap}{Table subcaptions. Set as Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#tables}{\code{tbl-subcap}} code chunk option. A
character vector.}

\item{layout_class}{Quarto \href{https://quarto.org/docs/authoring/article-layout.html#available-columns}{article layout class} for the figure output. Set as
Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#page-columns}{\code{tbl-column}} code chunk option. One of
\itemize{
\item \code{"body"}
\item \code{"body-outset"}
\item \code{"body-outset-left"}
\item \code{"body-outset-right"}
\item \code{"page"}
\item \code{"page-left"}
\item \code{"page-right"}
\item \code{"page-inset-left"}
\item \code{"page-inset-right"}
\item \code{"screen"}
\item \code{"screen-left"}
\item \code{"screen-right"}
\item \code{"screen-rightcolumn"}
\item \code{"screen-inset"}
\item \code{"screen-inset-shaded"}
\item \code{"screen-inset-left"}
\item \code{"screen-inset-right"}
\item \code{"margin"}
}}
}
\value{
A character scalar.
}
\description{
Assembles a \href{https://quarto.org/docs/authoring/tables.html#computations}{Quarto knitr table chunk}.
}
\details{
Use \link{substitute} together with \code{\link[=deparse1]{deparse1()}} to convert \R expressions to a character scalar:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{deparse1(expr = substitute(do_something()),
         collapse = "\\n")
}\if{html}{\out{</div>}}
}
\examples{
salim::quarto_tbl_chunk(body = "knitr::kable(head(cars))",
                        label = "tbl-head-cars",
                        tbl_cap = "Head of dataset `cars`",
                        layout_class = "margin") |>
  cat()
}
\seealso{
Other Quarto functions:
\code{\link{quarto_fig_chunk}()}
}
\concept{quarto}
