% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim-GEN.R
\name{rank_nr}
\alias{rank_nr}
\title{Convert an integer into spelled abbreviated English or German rank}
\usage{
rank_nr(x, lang = c("en", "de"))
}
\arguments{
\item{x}{An integer scalar or something coercible to.}

\item{lang}{The language to write the rank in.}
}
\value{
A character scalar.
}
\description{
Convert an integer into spelled abbreviated English or German rank
}
