% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{read_quarto_chunk_data}
\alias{read_quarto_chunk_data}
\title{Read in Quarto code chunk data}
\usage{
read_quarto_chunk_data(path)
}
\arguments{
\item{path}{Path to the TOML file containing structured figure or table code chunk data. A character scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Reads in a \href{https://toml.io/en/latest}{TOML} file containing structured figure or table code chunk data, suited to be fed to \code{\link[=quarto_chunks]{quarto_chunks()}}.
}
\seealso{
Other Quarto functions:
\code{\link{quarto_chunks}()},
\code{\link{quarto_fig_chunk}()},
\code{\link{quarto_tbl_chunk}()}
}
\concept{quarto}
