% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim-package.R
\docType{package}
\name{salim-package}
\alias{salim}
\alias{salim-package}
\title{salim: A Wild Mix of Functions Serving Various Purposes}
\description{
salim includes a diverse mix of functions which have not found their home in a more suitable place yet. The package kinda serves as an incubator for not yet fully matured code and thus will most probably never be made available through CRAN.
}
\section{Package configuration}{


Some of salim's functionality is controlled via package-specific global configuration which can either be set via \href{https://rdrr.io/r/base/options.html}{R options} or \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variables} (the former take precedence). This configuration includes:\tabular{llll}{
   \strong{Description} \tab \strong{R option} \tab \strong{Environment variable} \tab \strong{Default value} \cr
   Locale identifier consisting of a \href{https://en.wikipedia.org/wiki/ISO_639-1}{ISO 639-1} or \href{https://en.wikipedia.org/wiki/ISO_639-3}{ISO 639-3} language code plus an optional country identifier separated by a hyphen (\code{-}). For example \code{"fr"} or \code{"de-CH"}. \tab \code{salim.locale} \tab \code{R_SALIM_LOCALE} \tab \code{"en-US"} \cr
   Body font family for \href{https://salim.rpkg.dev/reference/#plot-theming}{plot theming functions}. \tab \code{salim.font_family_body} \tab \code{R_SALIM_FONT_FAMILY_BODY} \tab \code{"Alegreya Sans"} \cr
   Body color for \href{https://salim.rpkg.dev/reference/#plot-theming}{plot theming functions}. \tab \code{salim.plot_color_body} \tab \code{R_SALIM_PLOT_COLOR_BODY} \tab \code{"#343a40"} \cr
   Background color for \href{https://salim.rpkg.dev/reference/#plot-theming}{plot theming functions}. \tab \code{salim.plot_color_bg} \tab \code{R_SALIM_PLOT_COLOR_BG} \tab \code{"#f2f2f2"} \cr
   Grid color for \href{https://salim.rpkg.dev/reference/#plot-theming}{plot theming functions}. \tab \code{salim.plot_color_grid} \tab \code{R_SALIM_PLOT_COLOR_GRID} \tab \code{"#d9d9d9"} \cr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://salim.rpkg.dev}
  \item \url{https://gitlab.com/salim_b/r/pkgs/salim}
  \item Report bugs at \url{https://gitlab.com/salim_b/r/pkgs/salim/-/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
