% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{update_zdaarau_pkgs}
\alias{update_zdaarau_pkgs}
\title{Update R packages from the Centre for Democracy Studies Aarau (ZDA)}
\usage{
update_zdaarau_pkgs(pkgs = c("rdb.report", "fokus", "rdb", "swissevote"))
}
\arguments{
\item{pkgs}{R pkgs to be updated. A subset of:
\itemize{
\item \code{"rdb.report"}
\item \code{"fokus"}
\item \code{"rdb"}
\item \code{"swissevote"}
}}
}
\value{
\code{pkgs}, invisibly.
}
\description{
Installs/updates all of the \href{https://gitlab.com/zdaarau/rpkgs}{R packages from the Centre for Democracy Studies Aarau (ZDA)} to the latest development
version.
}
\seealso{
Other development environment currentness functions:
\code{\link{lvl_up_quarto}()},
\code{\link{lvl_up_rstudio}()},
\code{\link{lvl_up_r}()},
\code{\link{update_salims_pkgs}()}
}
\concept{dev_env}
