% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{write_out_n}
\alias{write_out_n}
\title{Write out a count}
\usage{
write_out_n(
  n,
  lang = c("en", "de"),
  use_singular = FALSE,
  use_dual = FALSE,
  strip_article = use_dual
)
}
\arguments{
\item{n}{The count to write out in letters. A non-negative integer scalar.}

\item{lang}{The language to write out \code{n} in.}

\item{use_singular}{Use a definite descriptor for \code{n = 0} and \code{n = 1} instead of the indistinct number. A logical scalar.}

\item{use_dual}{Use a definite descriptor for \code{n = 2} instead of the indistinct number. A logical scalar.}

\item{strip_article}{Strip the definite article from the definite descriptors. Only relevant if at least one of \code{use_singular} or \code{use_dual} is \code{TRUE}. A
logical scalar.}
}
\value{
A character scalar.
}
\description{
Write out a count
}
\seealso{
\code{\link[gt:vec_fmt_spelled_num]{gt::vec_fmt_spelled_num()}} which spells out integers from 0 to 100 in 80 different locales.

Other spoken language functions:
\code{\link{add_definite_article_de}()},
\code{\link{decline_noun_de}()},
\code{\link{definite_article_de_declined}()},
\code{\link{definite_article_de}()},
\code{\link{n_ify}()},
\code{\link{prettify_datetime}()},
\code{\link{prettify_date}()},
\code{\link{rank_nr}()}
}
\concept{spoken}
