% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim-GEN.R
\name{decline_noun_de}
\alias{decline_noun_de}
\title{Decline a German noun}
\usage{
decline_noun_de(
  noun,
  gender = c("feminine", "masculine", "neuter"),
  case = c("nominative", "genitive", "dative", "accusative")
)
}
\arguments{
\item{noun}{The noun to decline. A character scalar.}

\item{gender}{The grammatical gender of the \code{noun}.}

\item{case}{The grammatical case of the \code{noun}.}
}
\value{
A character scalar.
}
\description{
Note that this function currently doesn't support the \href{https://deutsch.lingolia.com/de/grammatik/nomen/deklination/genitiv#a-besonderheiten}{special cases}
where a masculine noun gets an \code{n} or \code{en} ending attached in the accusative, dative and genitive cases.
}
\seealso{
Other spoken language functions:
\code{\link{add_definite_article_de}()},
\code{\link{definite_article_de_declined}()},
\code{\link{definite_article_de}()},
\code{\link{rank_nr}()},
\code{\link{write_out_n}()}
}
\concept{spoken}
