% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{download_pandoc_binaries}
\alias{download_pandoc_binaries}
\title{Download Pandoc release binaries}
\usage{
download_pandoc_binaries(
  release_id = pandoc_release_id_latest(),
  os = c("linux", "macos", "windows"),
  path = "bin/",
  overwrite = TRUE
)
}
\arguments{
\item{release_id}{The GitHub release ID of the desired Pandoc release. Use \code{\link[=pandoc_releases]{pandoc_releases()}} to determine the release ID of a specific Pandoc version
number. An integer scalar.}

\item{os}{The operating systems for which Pandoc binaries should be downloaded. Any combination of
\code{"linux"},\code{"macos"} and \code{"windows"})`.}

\item{path}{The filesystem path to which the binaries are saved to. A \link[fs:fs_path]{path} or something coercible to.}

\item{overwrite}{Whether to overwrite existing binaries under \code{path}.}
}
\value{
\code{path} invisibly.
}
\description{
Downloads the executable program binaries of a certain Pandoc release for Linux, macOS and Windows.
}
