% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{lvl_up_rstudio}
\alias{lvl_up_rstudio}
\title{Level up RStudio}
\usage{
lvl_up_rstudio(path_min_vrsn, update_min_vrsn = FALSE)
}
\arguments{
\item{path_min_vrsn}{Path to the cached RStudio version string.}

\item{update_min_vrsn}{Whether or not to overwrite \code{path_min_vrsn} with the currently running RStudio version string \emph{iff} the latter is higher than the
former.}
}
\value{
Currently running RStudio version as a \link[=as.package_version]{numeric version}, or \code{NULL} if RStudio is not running, invisibly.
}
\description{
Checks whether the currently running RStudio version is >= a cached reference version. Intended to level up the RStudio version in use between multiple users
of the same code (e.g. contributors to a specific R project).
}
\details{
If the file \code{path_min_vrsn} exists, it is checked whether RStudio is running, and if so whether its version is greater than or equal to the version number
stored in that file, and if not, an alert is displayed.

If \code{update_min_vrsn} is set to \code{TRUE} and the file \code{path_min_vrsn} doesn't exist or contains an RStudio version string that's lower than the currently
running version of RStudio, \code{path_min_vrsn} is overwritten with the currently running RStudio version string.
}
\seealso{
Other development environment currentness functions:
\code{\link{lvl_up_r}()},
\code{\link{update_salims_pkgs}()}
}
\concept{dev_env}
