% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{pandoc_release_assets}
\alias{pandoc_release_assets}
\title{List Pandoc release assets}
\usage{
pandoc_release_assets(release_id = pandoc_release_id_latest())
}
\arguments{
\item{release_id}{The GitHub release ID of the desired Pandoc release. Use \code{\link[=pandoc_releases]{pandoc_releases()}} to determine the release ID of a specific Pandoc version
number. An integer scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Uses \code{\link[gh:gh]{gh::gh()}} to fetch filenames, corresponding operating systems and download URLs of a specific \href{https://docs.github.com/repositories/releasing-projects-on-github}{GitHub release} of \href{https://pandoc.org/}{Pandoc} via \href{https://docs.github.com/en/rest/reference/repos#list-release-assets}{GitHub's REST API} and returns them as a \link[tibble:tbl_df-class]{tibble}.
}
\seealso{
Other Pandoc-related functions:
\code{\link{download_pandoc_binaries}()},
\code{\link{pandoc_release_id_latest}()},
\code{\link{pandoc_releases}()},
\code{\link{pandoc_tpl}()},
\code{\link{pandoc_version_latest}()}
}
\concept{pandoc}
