% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{prettify_datetime}
\alias{prettify_datetime}
\title{Prettify datetime}
\usage{
prettify_datetime(datetime, locale = c("en", "de", "en-US", "de-CH"))
}
\arguments{
\item{datetime}{Datetime to be prettified. A \link[base:DateTimeClasses]{datetime} or something coercible to.}

\item{locale}{Locale the datetime should be prettified for. Currently only \code{"en"}/\code{"en-US"} and \code{"de"}/\code{"de-CH"} are implemented.}
}
\value{
A character scalar.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[stringi:stri_datetime_format]{stringi::stri_datetime_format()}} instead:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{stringi::stri_datetime_format(format = c("datetime_short", "datetime_medium", "datetime_long", "datetime_full"),
                              locale = "en")
}\if{html}{\out{</div>}}
}
\details{
Note that this might only work on (Ubuntu) Linux in the current form since locales are one bitchy hell of a PITA...
}
\examples{
salim:::prettify_datetime("2021-12-21T00:00:01Z")
}
\seealso{
Other spoken language functions:
\code{\link{add_definite_article_de}()},
\code{\link{decline_noun_de}()},
\code{\link{definite_article_de_declined}()},
\code{\link{definite_article_de}()},
\code{\link{n_ify}()},
\code{\link{prettify_date}()},
\code{\link{rank_nr}()},
\code{\link{write_out_n}()}
}
\concept{spoken}
\keyword{internal}
