% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{quarto_fig_chunk}
\alias{quarto_fig_chunk}
\title{Assemble Quarto knitr figure chunk}
\usage{
quarto_fig_chunk(
  body,
  label,
  fig_cap,
  fig_subcap = NULL,
  fig_column = "body",
  fig_width = NULL,
  fig_height = NULL,
  fig_pos = "H",
  fig_link = NULL,
  out_width = NULL,
  out_height = NULL
)
}
\arguments{
\item{body}{\R code to insert into the code chunk's body. A character scalar.}

\item{label}{Unique code chunk label. Set as Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#figures}{\code{label}} code chunk option. A
character scalar that starts with \code{"fig-"}.}

\item{fig_cap}{Figure caption. Set as Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#figures}{\code{fig-cap}} code chunk option. A character
scalar.}

\item{fig_subcap}{Figure subcaptions. Set as Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#figures}{\code{fig-subcap}} code chunk option. A
character vector, or \code{NULL} to omit.}

\item{fig_column}{Quarto \href{https://quarto.org/docs/authoring/article-layout.html#available-columns}{article layout class} for the figure output. Set as
Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#page-columns}{\code{fig-column}} code chunk option. One of
\itemize{
\item \code{"body"}
\item \code{"body-outset"}
\item \code{"body-outset-left"}
\item \code{"body-outset-right"}
\item \code{"page"}
\item \code{"page-left"}
\item \code{"page-right"}
\item \code{"page-inset-left"}
\item \code{"page-inset-right"}
\item \code{"screen"}
\item \code{"screen-left"}
\item \code{"screen-right"}
\item \code{"screen-inset"}
\item \code{"screen-inset-shaded"}
\item \code{"screen-inset-left"}
\item \code{"screen-inset-right"}
\item \code{"margin"}
}}

\item{fig_width}{Width of the plot (in inches), to be used in the graphics device. Set as Quarto's
\href{https://quarto.org/docs/reference/cells/cells-knitr.html#figures}{\code{fig-width}} code chunk option. A numeric scalar, or \code{NULL} to omit.}

\item{fig_height}{Height of the plot (in inches), to be used in the graphics device. Set as Quarto's
\href{https://quarto.org/docs/reference/cells/cells-knitr.html#figures}{\code{fig-height}} code chunk option. A numeric scalar, or \code{NULL} to omit.}

\item{fig_pos}{LaTeX figure position arrangement to be used in \verb{\\begin\{figure\}[]}. Set as Quarto's
\href{https://quarto.org/docs/reference/cells/cells-knitr.html#figures}{\code{fig-pos}} code chunk option. A character scalar, or \code{NULL} to omit. Use \code{"false"} for
no figure position specifier, which is sometimes necessary with custom figure environments (such as \code{sidewaysfigure}).}

\item{fig_link}{Hyperlink target for the figure. Set as Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#figures}{\code{fig-link}} code chunk
option. A character scalar, or \code{NULL} to omit.}

\item{out_width}{Width of the plot in the output document, which can be different from its physical \code{fig_width}, i.e., plots can be scaled in the output
document. Set as Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#figures}{\code{out-width}} code chunk option. A character scalar, or \code{NULL}
to omit.}

\item{out_height}{Height of the plot in the output document, which can be different from its physical \code{fig_height}, i.e., plots can be scaled in the output
document. Set as Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#figures}{\code{out-height}} code chunk option. A character scalar, or \code{NULL}
to omit.}
}
\value{
A character scalar.
}
\description{
Assembles a \href{https://quarto.org/docs/authoring/figures.html#computations}{Quarto knitr figure chunk}.
}
\details{
Use \link{substitute} together with \code{\link[=deparse1]{deparse1()}} to convert \R expressions to a character scalar as expected by param \code{body}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{deparse1(expr = substitute(do_something()),
         collapse = "\\n")
}\if{html}{\out{</div>}}
}
\examples{
salim::quarto_fig_chunk(body = "plot(cars)",
                        label = "fig-mtcars",
                        fig_cap = "Default plot for dataset `cars`",
                        fig_column = "page") |>
  cat()
}
\seealso{
Other Quarto functions:
\code{\link{quarto_chunks}()},
\code{\link{quarto_tbl_chunk}()},
\code{\link{read_quarto_chunk_data}()}
}
\concept{quarto}
