% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{quarto_tbl_chunk}
\alias{quarto_tbl_chunk}
\title{Assemble Quarto knitr table chunk}
\usage{
quarto_tbl_chunk(
  body,
  label,
  tbl_cap,
  tbl_subcap = NULL,
  tbl_colwidths = NULL,
  tbl_column = "body"
)
}
\arguments{
\item{body}{\R code to insert into the code chunk's body. A character scalar.}

\item{label}{Unique code chunk label. Set as Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#tables}{\code{label}} code chunk option. A
character scalar that starts with \code{"tbl-"}.}

\item{tbl_cap}{Table caption. Set as Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#tables}{\code{tbl-cap}} code chunk option. A character
scalar.}

\item{tbl_subcap}{Table subcaptions. Set as Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#tables}{\code{tbl-subcap}} code chunk option. A
character vector.}

\item{tbl_colwidths}{Apply explicit table column widths for markdown grid tables and pipe tables that are more than columns characters wide (72 by default).

Some formats (e.g. HTML) do an excellent job automatically sizing table columns and so don’t benefit much from column width specifications. Other formats
(e.g. LaTeX) require table column sizes in order to correctly flow longer cell content (this is a major reason why tables > 72 columns wide are assigned
explicit widths by Pandoc).

This can be specified as:
\itemize{
\item \code{"auto"}: Apply markdown table column widths except when there is a hyperlink in the table (which tends to throw off automatic calculation of column
widths based on the markdown text width of cells). \code{"auto"} is the default for HTML output formats.
\item \code{"true"}: Always apply markdown table widths. \code{"true"} is the default for all non-HTML formats.
\item \code{"false"}: Never apply markdown table widths.
\item A numeric vector (e.g. \code{c(40, 30, 30)}): Array of explicit width percentages.
}}

\item{tbl_column}{Quarto \href{https://quarto.org/docs/authoring/article-layout.html#available-columns}{article layout class} for the figure output. Set as
Quarto's \href{https://quarto.org/docs/reference/cells/cells-knitr.html#page-columns}{\code{tbl-column}} code chunk option. One of
\itemize{
\item \code{"body"}
\item \code{"body-outset"}
\item \code{"body-outset-left"}
\item \code{"body-outset-right"}
\item \code{"page"}
\item \code{"page-left"}
\item \code{"page-right"}
\item \code{"page-inset-left"}
\item \code{"page-inset-right"}
\item \code{"screen"}
\item \code{"screen-left"}
\item \code{"screen-right"}
\item \code{"screen-rightcolumn"}
\item \code{"screen-inset"}
\item \code{"screen-inset-shaded"}
\item \code{"screen-inset-left"}
\item \code{"screen-inset-right"}
\item \code{"margin"}
}}
}
\value{
A character scalar.
}
\description{
Assembles a \href{https://quarto.org/docs/authoring/tables.html#computations}{Quarto knitr table chunk}.
}
\details{
Use \link{substitute} together with \code{\link[=deparse1]{deparse1()}} to convert \R expressions to a character scalar as expected by param \code{body}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{deparse1(expr = substitute(do_something()),
         collapse = "\\n")
}\if{html}{\out{</div>}}
}
\examples{
salim::quarto_tbl_chunk(body = "knitr::kable(head(cars))",
                        label = "tbl-head-cars",
                        tbl_cap = "Head of dataset `cars`",
                        tbl_column = "margin") |>
  cat()
}
\seealso{
Other Quarto functions:
\code{\link{gen_quarto_snips}()},
\code{\link{quarto_fig_chunk}()},
\code{\link{read_quarto_chunk_data}()}
}
\concept{quarto}
