% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{rank_nr}
\alias{rank_nr}
\title{Convert an integer into spelled abbreviated English or German rank}
\usage{
rank_nr(x, lang = c("en", "de"))
}
\arguments{
\item{x}{An integer scalar or something coercible to.}

\item{lang}{The language to write the rank in.}
}
\value{
A character scalar.
}
\description{
Convert an integer into spelled abbreviated English or German rank
}
\seealso{
Other spoken language functions:
\code{\link{add_definite_article_de}()},
\code{\link{decline_noun_de}()},
\code{\link{definite_article_de_declined}()},
\code{\link{definite_article_de}()},
\code{\link{n_ify}()},
\code{\link{prettify_datetime}()},
\code{\link{prettify_date}()},
\code{\link{write_out_n}()}
}
\concept{spoken}
