% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\docType{data}
\name{regex_spelling_normalization}
\alias{regex_spelling_normalization}
\title{Regular expression patterns and replacements for spelling normalization}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
regex_spelling_normalization
}
\description{
Regular expression patterns and replacements for spelling normalization
}
\examples{
# unnest the pattern column
salim::regex_spelling_normalization |> tidyr::unnest_longer(col = pattern)
}
\seealso{
Other regular expression rules: \code{\link[yay:regex_text_normalization]{yay::regex_text_normalization}} \code{\link[yay:regex_file_normalization]{yay::regex_file_normalization}}
String normalization functions: \code{\link[yay:str_normalize]{yay::str_normalize()}} \code{\link[yay:str_normalize]{yay::str_normalize()}}
}
\keyword{datasets}
