% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim-package.R
\docType{package}
\name{salim-package}
\alias{salim}
\alias{salim-package}
\title{salim: A Wild Mix of Functions Serving Various Purposes}
\description{
salim includes a diverse mix of functions which have not
    found their home in a more suitable place yet. The package kinda
    serves as an incubator for not yet fully matured code and thus will
    most probably never be made available through CRAN.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/salim_b/r/pkgs/salim}
  \item Report bugs at \url{https://gitlab.com/salim_b/r/pkgs/salim/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
