% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{update_rpkgs}
\alias{update_rpkgs}
\title{Update R packages from rpkg.dev}
\usage{
update_rpkgs(
  pkgs = c("pal", "pkgpins", "pkgpurl", "pkgsnip", "plotlee", "qstnr", "rstd",
    "swissmuni", "tocr", "yay")
)
}
\arguments{
\item{pkgs}{R pkgs to be updated. A subset of:
\itemize{
\item \code{"pal"}
\item \code{"pkgpins"}
\item \code{"pkgpurl"}
\item \code{"pkgsnip"}
\item \code{"plotlee"}
\item \code{"qstnr"}
\item \code{"rstd"}
\item \code{"swissmuni"}
\item \code{"tocr"}
\item \code{"yay"}
}}
}
\value{
\code{pkgs}, invisibly.
}
\description{
Installs/updates \R packages from \href{https://gitlab.com/rpkg.dev}{rpkg.dev} to the latest development version.
}
\seealso{
Other development environment currentness functions:
\code{\link{lvl_up_quarto}()},
\code{\link{lvl_up_rstudio}()},
\code{\link{lvl_up_r}()},
\code{\link{update_salims_pkgs}()},
\code{\link{update_zdaarau_pkgs}()}
}
\concept{dev_env}
