% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{update_rpkgs}
\alias{update_rpkgs}
\title{Update R packages from rpkg.dev}
\usage{
update_rpkgs(
  pkgs = c("pal", "pkgpins", "pkgpurl", "pkgsnip", "plotlee", "qstnr", "rstd",
    "swissmuni", "tocr", "yay"),
  from_cran = FALSE
)
}
\arguments{
\item{pkgs}{R pkgs to be updated. A subset of:
\itemize{
\item \code{"pal"}
\item \code{"pkgpins"}
\item \code{"pkgpurl"}
\item \code{"pkgsnip"}
\item \code{"plotlee"}
\item \code{"qstnr"}
\item \code{"rstd"}
\item \code{"swissmuni"}
\item \code{"tocr"}
\item \code{"yay"}
}}

\item{from_cran}{Whether or not to install \code{pkgs} from a \href{https://en.wikipedia.org/wiki/R_package#Comprehensive_R_Archive_Network_(CRAN)}{CRAN} repository
if possible. The repository URLs set in the \code{\link[=options]{repos}} \R option are used by default, with fallback to \verb{https://cloud.r-project.org/}. If \code{FALSE},
the latest \emph{development} versions are always installed, regardless of whether \code{pkgs} are available on CRAN or not.}
}
\value{
\code{pkgs}, invisibly.
}
\description{
Installs/updates \R packages from \href{https://gitlab.com/rpkg.dev}{rpkg.dev} to the latest (development) version.
}
\seealso{
Other development environment currentness functions:
\code{\link{lvl_up_quarto}()},
\code{\link{lvl_up_r}()},
\code{\link{lvl_up_rstudio}()},
\code{\link{update_salims_pkgs}()},
\code{\link{update_zdaarau_pkgs}()}
}
\concept{dev_env}
