% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim-GEN.R
\name{update_salims_pkgs}
\alias{update_salims_pkgs}
\title{Update \href{https://gitlab.com/salim_b/r/pkgs}{Salim B's R packages}}
\usage{
update_salims_pkgs(
  pkgs = c("c2d4u", "pal", "pkgpins", "pkgpurl", "pkgsnip", "rstd", "salim",
    "swissmuni", "swissparty", "tocr", "yay")
)
}
\arguments{
\item{pkgs}{The R pkgs to be updated. A subset of:
\itemize{
\item \code{"c2d4u"}
\item \code{"pal"}
\item \code{"pkgpins"}
\item \code{"pkgpurl"}
\item \code{"pkgsnip"}
\item \code{"rstd"}
\item \code{"salim"}
\item \code{"swissmuni"}
\item \code{"swissparty"}
\item \code{"tocr"}
\item \code{"yay"} )`
}}
}
\value{
\code{pkgs}, invisibly.
}
\description{
Update \href{https://gitlab.com/salim_b/r/pkgs}{Salim B's R packages}
}
