% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salim.gen.R
\name{update_salims_pkgs}
\alias{update_salims_pkgs}
\title{Update Salim B's R packages}
\usage{
update_salims_pkgs(
  pkgs = c("c2d4u", "pal", "pkgpins", "pkgpurl", "pkgsnip", "qstnr", "rstd", "salim",
    "swissmuni", "swissparty", "tocr", "yay")
)
}
\arguments{
\item{pkgs}{R pkgs to be updated. A subset of:
\itemize{
\item \code{"c2d4u"}
\item \code{"pal"}
\item \code{"pkgpins"}
\item \code{"pkgpurl"}
\item \code{"pkgsnip"}
\item \code{"qstnr"}
\item \code{"rstd"}
\item \code{"salim"}
\item \code{"swissmuni"}
\item \code{"swissparty"}
\item \code{"tocr"}
\item \code{"yay"}
}}
}
\value{
\code{pkgs}, invisibly.
}
\description{
Installs/updates all of \href{https://gitlab.com/salim_b/r/pkgs}{Salim B's R packages} to the latest development version.
}
\seealso{
Other development environment currentness functions:
\code{\link{lvl_up_quarto}()},
\code{\link{lvl_up_rstudio}()},
\code{\link{lvl_up_r}()},
\code{\link{update_zdaarau_pkgs}()}
}
\concept{dev_env}
